chore: move truncate to shared CLI formatting utils file
